v0.7 test
Tweaked some of Player Info the layout.
We no longer draw a blank line for mana bar if we don't have one.
Shrank the size of the blink in combat bar and positioned it after your name.
Consolidated the player name class level onto one line.
Moved all % values to the center of their bars.
Tweaked some of Target layout.
Health Bar is easier to see.
Target information is drawn on top of the health bar.
Increased the percentage size and moved to the center of the bar.
Right Clicking an Icon will inspect the spell. Thank you Derple.
This does not work on EMU servers. maybe in the future.