Added border to buff icons for buffs that you are the caster of.  Currently this border is hard coded in MQ code to be gray, pending a change by brainiac to be pushed to the main branch which will allow it to be modified by ImStyleCol.Border
Added a RectFilled over buffs with 18 or less seconds remaining.  The alpha of this rectangle is modified to increase, or decrease (alternating) each time the targetBuffs function is ran.  This will give buffs with a low remaining duration a pulsing effect.

Modified ItemSpacing of the buffs to be 2x2 instead of 0x0 to prevent overlap of the borders on icons.

Modified maxIconRows to take the extra width from the ItemSpacing into account, as well as added a math.floor() around the calculation to always cause the number of icons per row to round down and not cause buffs to be cut off or drawn where the user is unable to see them.